+++ addOns/scripts/src/main/java/org/zaproxy/zap/extension/scripts/dialogs/LoadScriptDialog.java
@@ -32,7 +32,6 @@ import org.zaproxy.zap.extension.script.ScriptEngineWrapper;
 import org.zaproxy.zap.extension.script.ScriptType;
 import org.zaproxy.zap.extension.script.ScriptWrapper;
 import org.zaproxy.zap.extension.scripts.ExtensionScriptsUI;
-import org.zaproxy.zap.extension.scripts.NullScriptEngineWrapper;
 import org.zaproxy.zap.view.StandardFieldsDialog;
 
 public class LoadScriptDialog extends StandardFieldsDialog {
@@ -121,11 +120,6 @@ public class LoadScriptDialog extends StandardFieldsDialog {
     private List<String> getEngines() {
         ArrayList<String> list = new ArrayList<String>();
         list.addAll(extension.getExtScript().getScriptingEngines());
-        // TODO Remove the following workaround once targeting core version that respects
-        // ScriptEngineWrapper.isVisible.
-        // Remove the null scripting engine - unfortunately there is no easy way to do this
-        // and ExtensionScript.LANG_ENGINE_SEP (" : ") is private
-        list.remove(" : " + NullScriptEngineWrapper.NAME);
         return list;
     }